+++ src/prelude.rs
@@ -23,12 +23,12 @@ pub use toml;
 pub use crate::brush::*;
 pub use crate::config::*;
 pub use crate::definitions::*;
-pub use crate::loader::*;
-pub use crate::loader::bsp::*;
-pub use crate::loader::map::*;
+pub use crate::load::*;
+pub use crate::load::bsp::*;
+pub use crate::load::map::*;
 pub use crate::map_entity::*;
 pub use crate::material_properties::*;
-pub use crate::spawning::brushes::*;
-pub use crate::spawning::*;
+pub use crate::spawn::geometry::*;
+pub use crate::spawn::*;
 pub use crate::util::*;
 pub use crate::*;
